+++ tinted-builder/src/scheme/base16.rs
@@ -102,7 +102,7 @@ impl<'de> Deserialize<'de> for Base16Scheme {
                     )));
                 }
             }
-            _ => {
+            SchemeSystem::List | SchemeSystem::ListBase16 | SchemeSystem::ListBase24 => {
                 return Err(serde::de::Error::custom(format!(
                     "{} is not a valid Scheme system for a specific scheme",
                     wrapper.system